feat: heavy exif crud refactoring

+++ omoide/use_cases/api/uc_exif.py
@@ -4,7 +4,12 @@
 from uuid import UUID
 
 from omoide import domain
-from omoide.domain import interfaces, exceptions
+from omoide.domain import actions
+from omoide.domain import errors
+from omoide.domain import interfaces
+from omoide.infra.special_types import Failure
+from omoide.infra.special_types import Result
+from omoide.infra.special_types import Success
 from omoide.presentation import api_models
 
 __all__ = [
@@ -26,46 +31,34 @@ class BaseEXIFUseCase:
         self.items_repo = items_repo
         self.exif_repo = exif_repo
 
-    async def _assert_has_access(
-            self,
-            user: domain.User,
-            uuid: UUID,
-    ) -> domain.AccessStatus:
-        """Raise if user has no access to this EXIF."""
-        access = await self.items_repo.check_access(user, uuid)
-
-        if access.does_not_exist:
-            raise exceptions.NotFound(f'Item {uuid} does not exist')
-
-        if access.is_not_given:
-            raise exceptions.Forbidden(f'User {user.uuid} ({user.name}) '
-                                       f'has no access to item {uuid}')
-
-        if access.is_not_owner:
-            raise exceptions.Forbidden(f'You must own item {uuid} '
-                                       'to be able to modify it')
-
-        return access
-
 
 class CreateOrUpdateEXIFUseCase(BaseEXIFUseCase):
     """Use case for updating an EXIF."""
 
     async def execute(
             self,
+            policy: interfaces.AbsPolicy,
             user: domain.User,
             uuid: UUID,
             exif_in: api_models.EXIFIn,
-    ) -> bool:
+    ) -> Result[errors.Error, bool]:
         """Business logic."""
-        await self._assert_has_access(user, uuid)
+        async with self.items_repo.transaction():
+            error = await policy.is_restricted(user,
+                                               actions.EXIF.CREATE_OR_UPDATE)
 
-        exif = domain.EXIF(
-            item_uuid=uuid,
-            exif=exif_in.exif,
-        )
+            if error:
+                return Failure(error)
 
-        return await self.exif_repo.create_or_update_exif(user, exif)
+            exif = domain.EXIF(
+                item_uuid=uuid,
+                exif=exif_in.exif,
+                # TODO - crated_at? updated_at?
+            )
+
+            created = await self.exif_repo.create_or_update_exif(user, exif)
+
+        return Success(created)
 
 
 class ReadEXIFUseCase(BaseEXIFUseCase):
@@ -73,16 +66,22 @@ class ReadEXIFUseCase(BaseEXIFUseCase):
 
     async def execute(
             self,
+            policy: interfaces.AbsPolicy,
             user: domain.User,
             uuid: UUID,
-    ) -> domain.EXIF:
-        await self._assert_has_access(user, uuid)
-        exif = await self.exif_repo.read_exif(uuid)
+    ) -> Result[errors.Error, domain.EXIF]:
+        async with self.items_repo.transaction():
+            error = await policy.is_restricted(user, actions.EXIF.READ)
 
-        if exif is None:
-            raise exceptions.NotFound(f'EXIF {uuid} does not exist')
+            if error:
+                return Failure(error)
 
-        return exif
+            exif = await self.exif_repo.read_exif(uuid)
+
+            if exif is None:
+                return Failure(errors.EXIFDoesNotExist(uuid=uuid))
+
+        return Success(exif)
 
 
 class DeleteEXIFUseCase(BaseEXIFUseCase):
@@ -90,14 +89,20 @@ class DeleteEXIFUseCase(BaseEXIFUseCase):
 
     async def execute(
             self,
+            policy: interfaces.AbsPolicy,
             user: domain.User,
             uuid: UUID,
-    ) -> bool:
+    ) -> Result[errors.Error, bool]:
         """Business logic."""
-        await self._assert_has_access(user, uuid)
-        deleted = await self.exif_repo.delete_exif(uuid)
+        async with self.items_repo.transaction():
+            error = await policy.is_restricted(user, actions.EXIF.DELETE)
 
-        if not deleted:
-            raise exceptions.NotFound(f'EXIF {uuid} does not exist')
+            if error:
+                return Failure(error)
 
-        return True
+            deleted = await self.exif_repo.delete_exif(uuid)
+
+            if not deleted:
+                return Failure(errors.EXIFDoesNotExist(uuid=uuid))
+
+        return Success(True)
